// lib/api/generated/model/message.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * Wyng API
 * OpenAPI spec version: 1.0.0
 */
import type { MessageSender } from './messageSender';

export interface Message {
  id: string;
  matchId: string;
  senderId: string;
  body: string;
  isRead: boolean;
  createdAt: string;
  sender: MessageSender | null;
}